moved AccountPage component to components/
